done: refactor answerCheck function
- moved answerCheck to questionPage.js
- called updateScoreValue for realtime updating score
- removed extra check answer from updateScoreValue because check is already icluded in answerCheck
- removed utils folder
